VolumeDialog: BICPR changes to the slider
Okay, so what I did here is basically modifying the slider concept from @Extended-UI. If you want to preview it: https://imgur.com/a/aIQImg7

I set the border radius to fit the system, added caption button into it, changed the icon for settings and retinted the slider and the buttons. Enjoy!