Fix CDM.requestNotificationAccess() in profile.
If invoked by app running in managed profile, CompanionDeviceManager
.requestNotificationAccess() is actually trying to grant access to that
app in parent profile, which may not even exist.

Notification listener may not need to be running in managed profile,
but in order to access and control media sessions in managed profile,
notification listener must be granted access in that managed profile.

Test: Run test cases of CompanionDeviceManager in managed profile.

Signed-off-by: Oasis Feng <oasisfeng@gmail.com>
Change-Id: Id33c4956097300c5ec1c6c50ebee91dee9a49c33