frameworks: Apply our dark mode colour scheme
We previously set these as overlays, but noticed that on some builds this causes errors saying that the resource is not found. Ender told me it's related to the order it's being built at so when the overlay was built, the og resource wasn't even existing. So we now hardcode these values inside the frameworks to workaround this.